# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cucumber REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/feature": {
      "post": {
        "operationId": "createOrUpdateFeature",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feature"
              }
            }
          },
          "description": "The feature object to be created or updated",
          "required": true
        },
        "responses": {
          "default": {
            "content": {
              "*/*": {}
            },
            "description": "default response"
          }
        },
        "summary": "Create/update feature",
        "tags": [
          "Features"
        ]
      }
    },
    "/api/feature/{projectId}/{workItemId}": {
      "get": {
        "description": "Retrieve a feature by projectId and workItemId",
        "operationId": "getFeature",
        "parameters": [
          {
            "description": "ID of the project",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ID of the work item",
            "in": "path",
            "name": "workItemId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            },
            "description": "Successfully retrieved the feature"
          }
        },
        "summary": "Get feature",
        "tags": [
          "Features"
        ]
      }
    },
...
    "/raven/1.0/export/test": {
      "get": {
        "operationId": "exportTest",
        "parameters": [
          {
            "description": "String which contain list of work items information (e.g. 'elibrary/EL-103;drivepilot/DP-47')",
            "in": "query",
            "name": "keys",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Query which will be passed to the trackerService#queryWorkItems method",
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "True - the response will contain zip-archive with all workitems found, false - only first workitem with no archiving",
            "in": "query",
            "name": "fz",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the list of features"
          }
        },
        "summary": "Get features list",
        "tags": [
          "Export"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Feature": {
        "description": "Represents a feature, including project and work item details, as well as feature content",
        "properties": {
          "content": {
            "description": "The content of the feature",
            "type": "string"
          },
          "filename": {
            "description": "The name of the file containing the feature",
            "type": "string"
          },
          "projectId": {
            "description": "The ID of the project associated with the feature",
            "type": "string"
          },
          "title": {
            "description": "The title of the feature",
            "type": "string"
          },
          "workItemId": {
            "description": "The ID of the work item linked to the feature",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  }
}